(no-op) Move xdr_verify test-only parser wrappers into tests/utils.rs
